+++ src/main/java/model/HttpRequest.java
@@ -14,12 +14,9 @@ import java.util.List;
 import java.util.Map;
 
 public class HttpRequest {
-    private final Map<String, String> headerMap = new HashMap<>();
-    private final Map<String, String> queryParameterMap = new HashMap<>();
+    private HttpRequestLine httpRequestLine;
+    private HttpRequestHeader httpRequestHeader;
 
-    private String method;
-    private String path;
-    private String protocol;
     private String body = "";
 
     public HttpRequest(InputStream in) throws IOException {
@@ -28,86 +25,15 @@ public class HttpRequest {
 
     public HttpRequest(BufferedReader br) throws IOException {
         String line = br.readLine();
-        setStartLine(line);
-        setHeaderLine(br);
+        httpRequestLine = new HttpRequestLine(line);
+        httpRequestHeader = new HttpRequestHeader(br);
         setBody(br);
     }
 
     private void setBody(BufferedReader br) throws IOException {
         while(br.ready()){
-            body = IOUtils.readData(br, Integer.parseInt(headerMap.get("Content-Length")));
+            body = IOUtils.readData(br, Integer.parseInt(getHeader("Content-Length")));
         }
-    }
-
-    private void setHeaderLine(BufferedReader br) throws IOException {
-        String line = br.readLine();
-        while (!(line == null || line.trim().equals(""))) {
-            String[] headerLine = line.split(":", 2);
-            validateHeaderLine(headerLine);
-            headerMap.put(headerLine[0].trim(), headerLine[1].trim());
-            line = br.readLine();
-        }
-    }
-
-    private void validateHeaderLine(String[] headerLine) {
-        if (headerLine.length < 2) {
-            throw new IllegalHeaderException();
-        }
-    }
-
-    private void setStartLine(String line) {
-        String[] startLine = line.split(" ");
-        validateStartLine(startLine);
-        method = startLine[0];
-        setPath(startLine[1]);
-        protocol = startLine[2];
-    }
-
-    private void setPath(String path) {
-        String[] pathArgs = path.split("\\?", 2);
-        this.path = pathArgs[0];
-        if (pathArgs.length == 1) {
-            return;
-        }
-        setQueryParameters(pathArgs[1]);
-    }
-
-    private void setQueryParameters(String pathArg) {
-        String[] queryParams = pathArg.split("&");
-        Arrays.stream(queryParams).forEach((param) -> {
-            String[] args = param.split("=", 2);
-            queryParameterMap.put(args[0].trim(), args[1].trim());
-        });
-    }
-
-    private static void validateStartLine(String[] startLine) {
-        if (startLine.length != 3 || !HttpMethod.valueOf(startLine[0]).name().equals(startLine[0]) || !startLine[1].startsWith("/")) {
-            throw new IllegalHttpRequestException();
-        }
-    }
-
-    public String getMethod() {
-        return method;
-    }
-
-    public String getPath() {
-        return path;
-    }
-
-    public String getProtocol() {
-        return protocol;
-    }
-
-    public Map<String, String> getHeaderMap() {
-        return headerMap;
-    }
-
-    public String getHeader(String key){
-        String header = headerMap.get(key);
-        if(header == null) {
-            throw new IllegalHeaderException();
-        }
-        return header;
     }
 
     public String getBody() {
@@ -136,7 +62,27 @@ public class HttpRequest {
         return parsedCookie.get(cookieParam);
     }
 
+    public Map<String, String> getHeaderMap() {
+        return httpRequestHeader.getHeaderMap();
+    }
+
+    public String getHeader(String key){
+        return httpRequestHeader.getHeader(key);
+    }
+
+    public String getMethod() {
+        return httpRequestLine.getMethod();
+    }
+
+    public String getPath() {
+        return httpRequestLine.getPath();
+    }
+
+    public String getProtocol() {
+        return httpRequestLine.getProtocol();
+    }
+
     public Map<String, String> getQueryParameterMap() {
-        return queryParameterMap;
+        return httpRequestLine.getQueryParameterMap();
     }
 }
